Coerce to length-1 factor

Checks whether a vector can be coerced to a length-1 factor. to_fct_scalar() is optimized to check for length-1 factors (compared to stabilize_fct() with max_size = 1). to_factor_scalar() is a synonym of to_fct_scalar().

Usage

to_fct_scalar(
  x,
  ...,
  allow_null = FALSE,
  allow_zero_length = FALSE,
  levels = NULL,
  to_na = character(),
  x_arg = caller_arg(x),
  call = caller_env(),
  x_class = object_type(x)
)

to_factor_scalar(
  x,
  ...,
  allow_null = FALSE,
  allow_zero_length = FALSE,
  levels = NULL,
  to_na = character(),
  x_arg = caller_arg(x),
  call = caller_env(),
  x_class = object_type(x)
)

Arguments

x

The object to stabilize.

...

Arguments passed to methods.

allow_null

(logical(1)) Is NULL an acceptable value?

allow_zero_length

(logical(1)) Are zero-length vectors acceptable?

levels

(character) Expected levels. If NULL (default), the levels will be computed by base::factor().

to_na

(character) Values to convert to NA.

x_arg

(character(1)) The name of the object being stabilized to use in error messages. The automatic value will work in most cases, or pass it through from higher-level functions to make error messages clearer in unexported functions.

call

(environment) The execution environment to mention as the source of error messages.

x_class

(character(1)) The class name of the object being stabilized to use in error messages. Use this if you remove a special class from the object before checking its coercion, but want the error message to match the original class.

Value

The input as a length-1 factor, or an error condition with classes <stbl-error>, <stbl-condition>, <rlang_error>, <error>, <condition>, and a specific class by failure mode:

<stbl-error-coerce-factor> when x cannot be coerced to factor.

<stbl-error-incompatible_values-factor> when some elements of a list cannot be converted to factor.

<stbl-error-fct_levels> when values are not present in levels.

<stbl-error-bad_null> for NULL values when allow_null = FALSE.

<stbl-error-bad_empty> for empty vectors when allow_zero_length = FALSE.

<stbl-error-non_scalar> for non-scalar vectors.

See also

Examples

to_fct_scalar("a")
#> [1] a
#> Levels: a
try(to_fct_scalar(letters))
#> Error in eval(expr, envir) : 
#>   `letters` must be a single <factor>.
#> ✖ `letters` has 26 values.