Shared params for list helpers

Shared params for list helpers

Arguments

element_specs

(list) Named list of stabilizer functions, such as stabilize_* functions (stabilize_chr(), etc) or functions produced by specify_*() functions (specify_chr(), etc). Each name corresponds to a required element in .x, and the function is used to validate that element.

is_extra_named

(logical) Element-wise indicator of extra named positions.

named_spec

A single stabilizer function, such as a stabilize_* function (stabilize_chr(), etc) or a function produced by a specify_*() function (specify_chr(), etc), or NULL to disallow extra named elements.

nms

(character) Result of rlang::names2(.x).