Abort with a standardized "can't coerce" message

Abort with a standardized "can't coerce" message

Usage

.stop_cant_coerce(from_class, to_class, x_arg, call, additional_msg = NULL)

Arguments

from_class

(length-1 character) The class of the object that failed coercion.

to_class

(length-1 character) The target class for the coercion.

x_arg

(length-1 character) An argument name for x. The automatic value will work in most cases, or pass it through from higher-level functions to make error messages clearer in unexported functions.

call

The execution environment of the call. See the call argument of rlang::abort() for more information.

additional_msg

(length-1 character) Optional, additional cli-formatted messages.

Value

This function is called for its side effect of throwing an error and does not return a value.